refactor: restore 20-line headroom on ten near-cap app lib hosts (#4606)
Topic-split Slice C files that sat at 241–250 physical lines (including ungated settings_dialog) so each host and sibling is ≤230, keeping SettingsDialog.screenId and overlay factory names stable.
