+++ app/lib/features/game/flame/map_state/province_army_move_action_state.dart
@@ -3,7 +3,10 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 
 import '../caches/per_player_army_move_picker_cache.dart';
+import 'province_army_move_action_state_invade.dart';
 import 'province_army_move_home_army.dart';
+
+export 'province_army_move_action_state_invade.dart';
 
 /// Visibility/enablement for MAP20001 Military Move / Invade (Refs #4350).
 ///
@@ -80,7 +83,7 @@ ProvinceArmyMoveActionState computeProvinceArmyMoveActionState({
     );
   }
 
-  if (ownerId == null || !_hasInvasionOwnerSemantics(game, ownerId)) {
+  if (ownerId == null || !hasInvasionOwnerSemantics(game, ownerId)) {
     return ProvinceArmyMoveActionState.hidden;
   }
 
@@ -192,50 +195,4 @@ ProvinceArmyMoveActionState _moveStateForOwnedProvince({
     invadeDisabledReason: ProvinceArmyMoveDisabledReason.none,
     eligibleInvadeArmyIds: [],
   );
-}
-
-/// Cheap Invade visibility predicate (topology + army stationing only).
-bool invadeConceivableCheap({
-  required Game game,
-  required MapTopology topology,
-  required String humanPlayerId,
-  required String targetFullProvinceId,
-}) {
-  final regionId = ProvinceId.regionIdFrom(targetFullProvinceId);
-  final localId = ProvinceId.localIdFrom(targetFullProvinceId);
-  final fieldInRegion = <Army>[
-    for (final army in game.worldState.armies)
-      if (army.ownerId == humanPlayerId &&
-          !army.isHomeArmy &&
-          army.regimentUnitIds.isNotEmpty &&
-          ProvinceId.regionIdFrom(army.stationedProvinceId) == regionId)
-        army,
-  ];
-  final neighborLocals = neighborProvinceIdsInRegion(
-    topology,
-    regionId,
-    localId,
-  ).toSet();
-  for (final army in fieldInRegion) {
-    final hostLocal = ProvinceId.localIdFrom(army.stationedProvinceId);
-    if (neighborLocals.contains(hostLocal)) return true;
-  }
-  return homeArmyDetachInvadeCheap(
-    game: game,
-    topology: topology,
-    humanPlayerId: humanPlayerId,
-    targetFullProvinceId: targetFullProvinceId,
-  );
-}
-
-bool _hasInvasionOwnerSemantics(Game game, String ownerId) {
-  if (ownerId.isEmpty) return false;
-  if (game.playerById(ownerId) != null) return true;
-  for (final m in game.minorNations) {
-    if (m.id == ownerId) return true;
-  }
-  for (final t in game.tribes) {
-    if (t.id == ownerId) return true;
-  }
-  return false;
 }
