+++ app/lib/features/game/screens/production/production_screen_body.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_logic/industry_counsel_api.dart'
@@ -14,7 +12,6 @@ import 'package:colonizethis_turn/colonizethis_turn.dart'
 import 'package:colonizethis_world/colonizethis_world.dart';
 import 'package:colonizethis_app_fixtures/config/ct_e2e.dart';
 import 'package:colonizethis_app_fixtures/config/ct_e2e_last_panel_snapshot.dart';
-import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -24,13 +21,10 @@ import '../../../../core/services/game_service/try_get_game_map_data.dart';
 import '../../../../providers/game_service_provider.dart';
 import '../../../../providers/games_provider.dart';
 import '../../../../providers/production_allocation_provider.dart';
-import '../../widgets/production/production_commodity_breakdown_dialog.dart';
-import '../../widgets/production/production_labour_disband_confirm.dart';
-import '../../widgets/production/production_labour_helpers.dart';
-import '../../widgets/production/production_panel.dart';
 import '../../widgets/shell/shell_player_context.dart';
 import '../../widgets/shell/shell_player_guarded_body.dart';
 import 'production_screen.dart';
+import 'production_screen_body_panel.dart';
 
 /// Production screen body with shell guard and panel wiring (Refs #4117 de-part).
 class ProductionScreenBody extends ConsumerWidget {
@@ -156,75 +150,31 @@ class ProductionScreenBody extends ConsumerWidget {
       );
     }
 
-    final labourCallbacks = ProductionLabourCallbacks(
-      onAppendRecruitOrder: (tier) {
-        if (!canEdit) return;
-        final next = ordersWithAppendedRecruitWorkerOrder(
-          currentOrders: shellRef.read(currentOrdersProvider),
-          playerId: displayPlayer.id,
-          tier: tier,
-        );
-        shellRef.read(currentOrdersProvider.notifier).replaceAll(next);
-      },
-      onPopLastRecruitOrder: (tier) {
-        if (!canEdit) return;
-        final next = ordersWithLastRecruitWorkerOrderRemoved(
-          currentOrders: shellRef.read(currentOrdersProvider),
-          playerId: displayPlayer.id,
-          tier: tier,
-        );
-        shellRef.read(currentOrdersProvider.notifier).replaceAll(next);
-      },
-      onDisband: (tier) {
-        unawaited(
-          confirmAndApplyImmediateLabourDisband(
-            context: context,
-            tier: tier,
-            canEdit: canEdit,
-            readGame: () => shellRef.read(currentGameProvider) ?? displayGame,
-            writeGame: (nextGame) {
-              shellRef.read(currentGameProvider.notifier).setGame(nextGame);
-            },
-            playerId: displayPlayer.id,
-          ),
-        );
-      },
+    final labourCallbacks = productionScreenLabourCallbacks(
+      canEdit: canEdit,
+      playerId: displayPlayer.id,
+      shellRef: shellRef,
+      displayGame: displayGame,
+      context: context,
     );
-    final productionPanel = ProductionPanel(
-      game: displayGame,
-      player: displayPlayer,
+    final productionPanel = buildProductionScreenPanel(
+      context: context,
+      displayGame: displayGame,
+      displayPlayer: displayPlayer,
       desiredOutputByRecipe: desiredOutputByRecipe,
       netDeltasByCommodity: netDeltasByCommodity,
       labourReadiness: labourReadiness,
       forcesFeeding: forcesFeeding,
       currentOrders: currentOrders,
       labourCallbacks: labourCallbacks,
-      canEditLabour: canEdit,
-      onOpenCommodityBreakdown: canEdit
-          ? () {
-              showDialog<void>(
-                context: context,
-                barrierColor: EditorialMonoclePalette.dialogScrim,
-                builder: (_) => ProductionCommodityBreakdownDialog(
-                  game: displayGame,
-                  player: displayPlayer,
-                  topology: panelTopology,
-                  tileMapByRegion: panelTileMaps,
-                  currentOrders: currentOrders,
-                ),
-              );
-            }
-          : null,
-      onDesiredOutputChanged: (next) {
-        if (!canEdit) return;
-        shellRef
-            .read(productionDesiredOutputProvider.notifier)
-            .replaceAll(next);
-      },
+      canEdit: canEdit,
+      shellRef: shellRef,
+      panelTopology: panelTopology,
+      panelTileMaps: panelTileMaps,
       starredProduceRecommendationsByRecipeId:
           starredProduceRecommendationsByRecipeId,
-      onOpenCounsel: openCounsel,
-      onOpenTradeMarket: openTradeMarket,
+      openCounsel: openCounsel,
+      openTradeMarket: openTradeMarket,
     );
     if (kCtE2EEnabled) {
       updateCtE2eProductionPanelSnapshotIfEnabled(
