+++ app/lib/features/game/widgets/province_overlay/province_sea_zone_detail_overlay_province_content.dart
@@ -15,27 +15,13 @@ import 'package:colonizethis_app/features/game/flame/map_state/province_naval_mi
     show ProvinceNavalMissionOverlayControls;
 import 'package:colonizethis_app/features/game/flame/overlays/province_blockade_status_support.dart'
     show ProvinceBlockadeStatus;
-import 'package:colonizethis_app/core/utils/prefixed_id.dart';
-
-import 'province_overlay_unit_partition.dart';
-import 'province_sea_zone_detail_overlay_designation.dart';
-import 'province_sea_zone_detail_overlay_province_content_intel.dart';
-import 'province_sea_zone_detail_overlay_province_content_tabs.dart';
-import 'province_sea_zone_detail_overlay_province_content_unit_sections.dart';
-import 'province_sea_zone_detail_overlay_province_content_unrevealed.dart';
-import 'province_sea_zone_detail_overlay_sections_political.dart';
-import 'province_sea_zone_detail_overlay_support.dart';
-import 'province_sea_zone_detail_overlay_tile_section.dart';
-import 'package:colonizethis_app/features/game/flame/controls/map_tile_sight.dart';
-import 'package:colonizethis_app/features/game/widgets/units/civilian/work_order_afford_preview_ui.dart';
 import 'package:colonizethis_economy/colonizethis_economy.dart'
     show ProvinceImprovableCommodityCount;
-import 'package:colonizethis_world/colonizethis_world.dart'
-    show
-        PlayerView,
-        fleetsInPortAtProvince,
-        kRegionNewWorld,
-        provincePanelShowsFullTileDerivedIntel;
+import 'package:colonizethis_world/colonizethis_world.dart' show PlayerView;
+
+import 'province_sea_zone_detail_overlay_province_content_revealed.dart';
+import 'province_sea_zone_detail_overlay_province_content_unrevealed.dart';
+import 'province_sea_zone_detail_overlay_support.dart';
 
 OverlayContent provinceContent({
   required BuildContext context,
@@ -92,7 +78,6 @@ OverlayContent provinceContent({
   void Function(Iterable<String>?)? onHighlightTiles,
   ProvinceTileConnectivityDisplay? tileConnectivity,
 }) {
-  final regionId = prefixedIdRegionSegment(provinceId) ?? region.regionId;
   if (provinceContentIsFullyUnrevealed(
     region: region,
     provinceId: provinceId,
@@ -100,44 +85,7 @@ OverlayContent provinceContent({
   )) {
     return provinceContentUnrevealed(l10n: l10n);
   }
-  final province = findProvinceForSeaZoneOverlay(game, provinceId);
-  final regionData = provinceId.startsWith(kRegionNewWorld)
-      ? game.worldState.newWorld
-      : game.worldState.oldWorld;
-  final partitioned = partitionProvinceOverlayUnits(
-    regionUnits: regionData.units,
-    provinceId: provinceId,
-    humanPlayerId: humanPlayerId,
-    playerView: playerView,
-  );
-  final military = partitioned.military;
-  final civilian = partitioned.civilian;
-  final visibleCivilianCount = partitioned.visibleCivilianCount;
-  final fleetsInPort = fleetsInPortAtProvince(game.worldState, provinceId);
-  final tileKeys =
-      game.worldState.tileKeysByRegionAndProvince[region
-          .regionId]?[provinceId] ??
-      [];
-  final showsFullIntel =
-      omniscientDetail ||
-      provincePanelShowsFullTileDerivedIntel(
-        game: game,
-        view: playerView,
-        humanPlayerId: humanPlayerId,
-        provinceId: provinceId,
-        provinceTileKeys: tileKeys,
-      );
-  final tileIntel = aggregateProvinceTileIntel(
-    l10n: l10n,
-    game: game,
-    region: region,
-    provinceId: provinceId,
-    humanPlayerId: humanPlayerId,
-    playerView: playerView,
-    tileKeys: tileKeys,
-    omniscientDetail: omniscientDetail,
-  );
-  final tileSection = buildTileSection(
+  return provinceContentRevealed(
     context: context,
     l10n: l10n,
     game: game,
@@ -145,41 +93,29 @@ OverlayContent provinceContent({
     provinceId: provinceId,
     humanPlayerId: humanPlayerId,
     playerView: playerView,
-    civilianCount: visibleCivilianCount,
+    draftOrders: draftOrders,
     selectedTileKey: selectedTileKey,
+    onHighlightTile: onHighlightTile,
     civilianInlineActions: civilianInlineActions,
     inlineActionCallbacks: inlineActionCallbacks,
-    currentOrders: draftOrders,
-    tileConnectivity: tileConnectivity,
-    blockadeStatus: blockadeStatus,
-  );
-  final political = buildPoliticalSection(
-    l10n: l10n,
-    name: province?.displayName ?? provinceId,
-    ownerName: ownerNameForProvinceOverlay(l10n, game, province?.ownerId),
-    sightPhrase: mapTileSightPhraseForSelectedTile(
-      l10n: l10n,
-      region: region,
-      selectedTileKey: selectedTileKey,
-    ),
-    regionLabel: provinceOverlayRegionLabel(l10n, regionId),
-    isCapital: provinceOverlayIsCapital(game, provinceId),
-    townDevelopmentLevel:
-        province?.townDevelopmentLevel ?? kTownDevelopmentLevelMin,
     showUpgradeTownControl: showUpgradeTownControl,
     upgradeTownEnabled: upgradeTownEnabled,
-    upgradeTownTooltip: upgradeTownTargetTileKey == null
-        ? ''
-        : provinceOverlayPoliticalUpgradeTownTooltip(
-            l10n: l10n,
-            game: game,
-            humanPlayerId: humanPlayerId,
-            currentOrders: draftOrders,
-            townTileKey: upgradeTownTargetTileKey,
-            enabled: upgradeTownEnabled,
-            hasBuilderUnits: upgradeTownHasBuilderUnits,
-          ),
+    upgradeTownHasBuilderUnits: upgradeTownHasBuilderUnits,
+    upgradeTownTargetTileKey: upgradeTownTargetTileKey,
     onUpgradeTownTap: onUpgradeTownTap,
+    showMoveArmyControl: showMoveArmyControl,
+    moveArmyEnabled: moveArmyEnabled,
+    moveArmyTooltip: moveArmyTooltip,
+    onMoveArmyTap: onMoveArmyTap,
+    showInvadeArmyControl: showInvadeArmyControl,
+    invadeArmyEnabled: invadeArmyEnabled,
+    invadeArmyTooltip: invadeArmyTooltip,
+    onInvadeArmyTap: onInvadeArmyTap,
+    navalMission: navalMission,
+    detachAndSail: detachAndSail,
+    stationSpy: stationSpy,
+    counterEspionage: counterEspionage,
+    blockadeStatus: blockadeStatus,
     showEstablishConsulateControl: showEstablishConsulateControl,
     establishConsulateEnabled: establishConsulateEnabled,
     establishConsulatePending: establishConsulatePending,
@@ -194,53 +130,11 @@ OverlayContent provinceContent({
     offerPeaceRejectionReason: offerPeaceRejectionReason,
     onOfferPeaceTap: onOfferPeaceTap,
     isNarrow: isNarrow,
-  );
-  final buildFort = civilianInlineActions.buildFort;
-  final unitSections = buildProvinceIntelGatedUnitSections(
-    l10n: l10n,
-    game: game,
-    showsFullIntel: showsFullIntel,
-    humanPlayerId: humanPlayerId,
-    provinceId: provinceId,
-    draftOrders: draftOrders,
-    playerView: playerView,
-    military: military,
-    civilian: civilian,
-    fleetsInPort: fleetsInPort,
-    fortLevel: province?.fortLevel ?? 0,
-    buildFortAction: buildFort,
-    onBuildFortTap: inlineActionCallbacks.onBuildFortTap,
-    showMoveArmyControl: showMoveArmyControl,
-    moveArmyEnabled: moveArmyEnabled,
-    moveArmyTooltip: moveArmyTooltip,
-    onMoveArmyTap: onMoveArmyTap,
-    showInvadeArmyControl: showInvadeArmyControl,
-    invadeArmyEnabled: invadeArmyEnabled,
-    invadeArmyTooltip: invadeArmyTooltip,
-    onInvadeArmyTap: onInvadeArmyTap,
-    navalMission: navalMission,
-    detachAndSail: detachAndSail,
-    blockadeStatus: blockadeStatus,
-    stationSpy: stationSpy,
-    counterEspionage: counterEspionage,
-    provinceDisplayName: province?.displayName,
-    onHighlightTile: onHighlightTile,
-    onHighlightTiles: onHighlightTiles,
+    omniscientDetail: omniscientDetail,
+    townProductionBonusByCommodity: townProductionBonusByCommodity,
     extractionSnapshot: extractionSnapshot,
     availableByCommodity: availableByCommodity,
-    townProductionBonusByCommodity: townProductionBonusByCommodity,
-    byResImproved: tileIntel.byResImproved,
-    byResImprovable: tileIntel.byResImprovable,
-    resourceKeysSorted: tileIntel.resourceKeysSorted,
-    selectedTileKey: selectedTileKey,
-  );
-  return assembleProvinceOverlayTabContent(
-    l10n: l10n,
-    political: political,
-    tileSection: tileSection,
-    economic: unitSections.economic,
-    militarySection: unitSections.military,
-    civilianSection: unitSections.civilian,
-    naval: unitSections.naval,
+    onHighlightTiles: onHighlightTiles,
+    tileConnectivity: tileConnectivity,
   );
 }
