+++ app/test/features/shell/settings/settings_dialog_test.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:colonizethis_app/config/constants.dart';
 import 'package:colonizethis_app/features/game/flame/map_theme/map_theme_catalog_loader.dart';
 import 'package:colonizethis_app/features/game/flame/map_theme/map_theme_models.dart';
+import 'package:colonizethis_app/config/ui_screen_ids.dart';
 import 'package:colonizethis_app/features/shell/settings/settings_dialog.dart';
 import 'package:colonizethis_app/providers/settings_provider.dart';
 import 'package:colonizethis_app/config/ux_settings_keys.dart';
@@ -24,22 +25,25 @@ void main() {
     MapThemeCatalogLoader.resetForTest();
   });
 
+  test('SettingsDialog.screenId stays UiScreenIds.settingsDialog', () {
+    expect(SettingsDialog.screenId, UiScreenIds.settingsDialog);
+  });
+
   testWidgets('Settings dialog shows multi-theme pickers only', (tester) async {
     await MapThemeCatalogLoader.ensureLoaded();
 
     await tester.pumpWidget(
       buildAppShell(
-        child: const SizedBox(
-          width: 400,
-          height: 600,
-          child: SettingsDialog(),
-        ),
+        child: const SizedBox(width: 400, height: 600, child: SettingsDialog()),
       ),
     );
     await tester.pump();
 
     expect(find.byKey(SettingsDialog.titleKey), findsOneWidget);
-    expect(find.byKey(SettingsDialog.warnIdleCiviliansToggleKey), findsOneWidget);
+    expect(
+      find.byKey(SettingsDialog.warnIdleCiviliansToggleKey),
+      findsOneWidget,
+    );
     expect(find.byType(CtToggleSwitch), findsOneWidget);
     expect(find.byType(CtDropdown<String>), findsNWidgets(2));
     expect(
@@ -76,41 +80,40 @@ void main() {
           .read(settingsProvider.notifier)
           .setValue(UxSettingsKeys.warnIdleCiviliansOnEndTurn, false);
       expect(
-        container.read(settingsProvider)[UxSettingsKeys.warnIdleCiviliansOnEndTurn],
+        container.read(
+          settingsProvider,
+        )[UxSettingsKeys.warnIdleCiviliansOnEndTurn],
         isFalse,
       );
       expect(box.get(UxSettingsKeys.warnIdleCiviliansOnEndTurn), isFalse);
     },
   );
 
-  test(
-    'settingsProvider setValue persists mapTheme.terrain in Hive',
-    () async {
-      final dir = await Directory.systemTemp.createTemp('ct_settings_dlg_');
-      Hive.init(dir.path);
-      await Hive.openBox<dynamic>(HiveBoxNames.settings);
-      addTearDown(() async {
-        await Hive.close();
-        await dir.delete(recursive: true);
-      });
+  test('settingsProvider setValue persists mapTheme.terrain in Hive', () async {
+    final dir = await Directory.systemTemp.createTemp('ct_settings_dlg_');
+    Hive.init(dir.path);
+    await Hive.openBox<dynamic>(HiveBoxNames.settings);
+    addTearDown(() async {
+      await Hive.close();
+      await dir.delete(recursive: true);
+    });
 
-      final container = ProviderContainer();
-      addTearDown(container.dispose);
+    final container = ProviderContainer();
+    addTearDown(container.dispose);
 
-      container
-          .read(settingsProvider.notifier)
-          .setValue(MapThemeGroupId.terrain.settingsKey, 'sepia');
+    container
+        .read(settingsProvider.notifier)
+        .setValue(MapThemeGroupId.terrain.settingsKey, 'sepia');
 
-      expect(
-        container.read(settingsProvider)[MapThemeGroupId.terrain.settingsKey],
-        'sepia',
-      );
-      expect(
-        Hive.box<dynamic>(
-          HiveBoxNames.settings,
-        ).get(MapThemeGroupId.terrain.settingsKey),
-        'sepia',
-      );
-    },
-  );
+    expect(
+      container.read(settingsProvider)[MapThemeGroupId.terrain.settingsKey],
+      'sepia',
+    );
+    expect(
+      Hive.box<dynamic>(
+        HiveBoxNames.settings,
+      ).get(MapThemeGroupId.terrain.settingsKey),
+      'sepia',
+    );
+  });
 }
